+++ targeting/conditions/conversion_condition.go
@@ -12,31 +12,37 @@ func NewConversionCondition(c types.TargetingCondition) *ConversionCondition {
 			Type:    c.Type,
 			Include: c.Include,
 		},
-		GoalId: c.GoalId,
+		visitorScopeCondition: newVisitorScopeCondition(c, VisitScopeVisitor),
+		GoalId:                c.GoalId,
 	}
 }
 
 type ConversionCondition struct {
 	types.TargetingConditionBase
+	visitorScopeCondition
 	GoalId int `json:"goalId"`
 }
 
 func (c *ConversionCondition) CheckTargeting(targetData interface{}) bool {
-	conversionStorage, ok := targetData.(storage.DataCollectionStorage[*types.Conversion])
-	if ok && (conversionStorage != nil) {
-		if c.GoalId == 0 && conversionStorage.Len() > 0 {
-			return true
-		}
-		targeted := false
-		conversionStorage.Enumerate(func(conversion *types.Conversion) bool {
-			targeted = c.GoalId == conversion.GoalId()
-			return !targeted
-		})
-		return targeted
+	targetingData, ok := targetData.(TargetingDataConversionCondition)
+	if !ok || (targetingData.Conversions == nil) || (targetingData.Conversions.Len() == 0) {
+		return false
 	}
-	return false
+	assignmentThresholdMillis := c.assignmentThresholdMillis(targetingData.VisitorVisits)
+	targeted := false
+	targetingData.Conversions.Enumerate(func(conversion *types.Conversion) bool {
+		targeted = ((c.GoalId == types.UndefinedGoalId) || (c.GoalId == conversion.GoalId())) &&
+			(conversion.AssignmentTime().UnixMilli() >= assignmentThresholdMillis)
+		return !targeted
+	})
+	return targeted
 }
 
 func (c ConversionCondition) String() string {
 	return utils.JsonToString(c)
+}
+
+type TargetingDataConversionCondition struct {
+	VisitorVisits *types.VisitorVisits
+	Conversions   storage.DataCollectionStorage[*types.Conversion]
 }