+++ targeting/conditions/exclusive_experiment.go
@@ -15,6 +15,7 @@ const (
 
 type ExclusiveExperimentCondition struct {
 	types.TargetingConditionBase
+	visitorScopeCondition
 	campaignType string
 }
 
@@ -24,35 +25,59 @@ func NewExclusiveExperimentCondition(c types.TargetingCondition) *ExclusiveExper
 			Type:    c.Type,
 			Include: true,
 		},
-		campaignType: c.CampaignType,
+		visitorScopeCondition: newVisitorScopeCondition(c, VisitScopeVisitor),
+		campaignType:          c.CampaignType,
 	}
 }
 
 func (c *ExclusiveExperimentCondition) CheckTargeting(targetData interface{}) bool {
 	if targetingData, ok := targetData.(TargetingDataExclusiveExperiment); ok {
+		assignmentThresholdMillis := c.assignmentThresholdMillis(targetingData.VisitorVisits)
 		switch c.campaignType {
 		case CampaignTypeExperiment:
-			return c.checkExperiment(targetingData)
+			return c.checkExperiment(targetingData, assignmentThresholdMillis)
 		case CampaignTypePersonalization:
-			return c.checkPersonalization(targetingData)
+			return c.checkPersonalization(targetingData, assignmentThresholdMillis)
 		case CampaignTypeAny:
-			return c.checkPersonalization(targetingData) && c.checkExperiment(targetingData)
+			return c.checkPersonalization(targetingData, assignmentThresholdMillis) &&
+				c.checkExperiment(targetingData, assignmentThresholdMillis)
 		}
 		logging.Error("Unexpected campaign type for %s condition: %s", c.Type, c.campaignType)
 	}
 	return false
 }
 
-func (*ExclusiveExperimentCondition) checkExperiment(td TargetingDataExclusiveExperiment) bool {
+func (*ExclusiveExperimentCondition) checkExperiment(
+	td TargetingDataExclusiveExperiment, assignmentThresholdMillis int64,
+) bool {
 	if td.Variations == nil {
 		return true
 	}
-	variationCount := td.Variations.Len()
-	return (variationCount == 0) || ((variationCount == 1) && (td.Variations.Get(td.CurrentExperimentId) != nil))
+	targeted := true
+	td.Variations.Enumerate(func(assignedVariation *types.AssignedVariation) bool {
+		if (assignedVariation.ExperimentId() != td.CurrentExperimentId) &&
+			(assignedVariation.AssignmentTime().UnixMilli() >= assignmentThresholdMillis) {
+			targeted = false
+		}
+		return targeted
+	})
+	return targeted
 }
 
-func (*ExclusiveExperimentCondition) checkPersonalization(td TargetingDataExclusiveExperiment) bool {
-	return (td.Personalizations == nil) || (td.Personalizations.Len() == 0)
+func (*ExclusiveExperimentCondition) checkPersonalization(
+	td TargetingDataExclusiveExperiment, assignmentThresholdMillis int64,
+) bool {
+	if td.Personalizations == nil {
+		return true
+	}
+	targeted := true
+	td.Personalizations.Enumerate(func(personalization *types.Personalization) bool {
+		if personalization.AssignmentTime().UnixMilli() >= assignmentThresholdMillis {
+			targeted = false
+		}
+		return targeted
+	})
+	return targeted
 }
 
 func (c ExclusiveExperimentCondition) String() string {
@@ -60,6 +85,7 @@ func (c ExclusiveExperimentCondition) String() string {
 }
 
 type TargetingDataExclusiveExperiment struct {
+	VisitorVisits       *types.VisitorVisits
 	CurrentExperimentId int
 	Variations          storage.DataMapStorage[int, *types.AssignedVariation]
 	Personalizations    storage.DataMapStorage[int, *types.Personalization]