+++ targeting/conditions/target_experiment.go
@@ -8,6 +8,7 @@ import (
 
 type TargetExperimentCondition struct {
 	types.TargetingConditionBase
+	visitorScopeCondition
 	variationId        int
 	experimentId       int
 	variationMatchType types.OperatorType
@@ -19,22 +20,34 @@ func NewTargetExperimentCondition(c types.TargetingCondition) *TargetExperimentC
 			Type:    c.Type,
 			Include: true,
 		},
-		variationId:        c.VariationId,
-		experimentId:       c.ExperimentId,
-		variationMatchType: c.VariationMatchType,
+		visitorScopeCondition: newVisitorScopeCondition(c, VisitScopeCurrentVisit),
+		variationId:           c.VariationId,
+		experimentId:          c.ExperimentId,
+		variationMatchType:    c.VariationMatchType,
 	}
 }
 
 func (c *TargetExperimentCondition) CheckTargeting(targetData interface{}) bool {
-	if variations, ok := targetData.(storage.DataMapStorage[int, *types.AssignedVariation]); ok {
-		variation := variations.Get(c.experimentId)
-		switch c.variationMatchType {
-		case types.OperatorAny:
-			return variation != nil
-		case types.OperatorExact:
-			return (variation != nil) && (variation.VariationId() == c.variationId)
-		}
-		logging.Error("Unexpected variation match type for %s condition: %s", c.Type, c.variationMatchType)
+	targetingData, ok := targetData.(TargetingDataTargetExperimentCondition)
+	if !ok || (targetingData.VariationStorage == nil) {
+		return false
 	}
+	variation := targetingData.VariationStorage.Get(c.experimentId)
+	if (variation != nil) &&
+		(variation.AssignmentTime().UnixMilli() < c.assignmentThresholdMillis(targetingData.VisitorVisits)) {
+		variation = nil
+	}
+	switch c.variationMatchType {
+	case types.OperatorAny:
+		return variation != nil
+	case types.OperatorExact:
+		return (variation != nil) && (variation.VariationId() == c.variationId)
+	}
+	logging.Error("Unexpected variation match type for %s condition: %s", c.Type, c.variationMatchType)
 	return false
+}
+
+type TargetingDataTargetExperimentCondition struct {
+	VisitorVisits    *types.VisitorVisits
+	VariationStorage storage.DataMapStorage[int, *types.AssignedVariation]
 }