+++ types/personalization.go
@@ -1,14 +1,19 @@
 package types
 
-import "fmt"
+import (
+	"fmt"
+	"time"
+)
 
 type Personalization struct {
-	id          int
-	variationId int
+	id             int
+	variationId    int
+	assignmentTime time.Time
 }
 
-func NewPersonalization(id int, variationId int) *Personalization {
-	return &Personalization{id: id, variationId: variationId}
+// Note: This is intended for internal use only and is not part of the stable API.
+func NewPersonalization(id int, variationId int, assignmentTime time.Time) *Personalization {
+	return &Personalization{id: id, variationId: variationId, assignmentTime: assignmentTime}
 }
 
 func (p *Personalization) Id() int {
@@ -19,10 +24,19 @@ func (p *Personalization) VariationId() int {
 	return p.variationId
 }
 
+func (p *Personalization) AssignmentTime() time.Time {
+	return p.assignmentTime
+}
+
 func (*Personalization) DataType() DataType {
 	return DataTypePersonalization
 }
 
 func (p Personalization) String() string {
-	return fmt.Sprintf("Personalization{id:%d,variationId:%d}", p.id, p.variationId)
+	return fmt.Sprintf(
+		"Personalization{id:%d,variationId:%d,assignmentTime:%v}",
+		p.id,
+		p.variationId,
+		p.assignmentTime,
+	)
 }